// src/pages/settings/lib/types.ts
export type SettingsData = {
  id: number;
  firstName: string;
  lastName: string;
  status: string;
  profileImg: string;
  bannerImg: string;
  email: string;
};

export type IFormData = {
  firstName: string;
  lastName: string;
  status: string;
  profileImg: string;
  bannerImg: string;
  email: string;
};